+++ src/implementations/Agave/AgaveMultiStake.sol
@@ -108,7 +108,7 @@ contract AgaveMultiStake is StrategyV5 {
    * @param _index Index of the input to stake
    * @param _amount Amount of underlying assets to allocate to `inputs[_index]`
    */
-  function _stake(uint8 _index, uint256 _amount) internal override {
+  function _stake(uint256 _index, uint256 _amount) internal override {
     IPool pool = IPool(_poolProvider.getLendingPool());
     pool.deposit({
       asset: address(inputs[_index]),
@@ -123,7 +123,7 @@ contract AgaveMultiStake is StrategyV5 {
    * @param _index Index of the input to liquidate
    * @param _amount Amount of underlying assets to recover from liquidating `inputs[_index]`
    */
-  function _unstake(uint8 _index, uint256 _amount) internal override {
+  function _unstake(uint256 _index, uint256 _amount) internal override {
     IPool pool = IPool(_poolProvider.getLendingPool());
     pool.withdraw({asset: address(inputs[_index]), amount: _amount, to: address(this)});
   }
