+++ src/implementations/Toros/TorosMultiStake.sol
@@ -42,7 +42,7 @@ contract TorosMultiStake is StrategyV5 {
    * @param _index Index of the input to stake
    * @param _amount Amount of underlying assets to allocate to `inputs[_index]`
    */
-  function _stake(uint8 _index, uint256 _amount) internal override {
+  function _stake(uint256 _index, uint256 _amount) internal override {
     _dHedgeSwapper.deposit({
       pool: address(lpTokens[_index]),
       depositAsset: address(inputs[_index]),
@@ -57,7 +57,7 @@ contract TorosMultiStake is StrategyV5 {
    * @param _index Index of the input to liquidate
    * @param _amount Amount of underlying assets to recover from liquidating `inputs[_index]`
    */
-  function _unstake(uint8 _index, uint256 _amount) internal override {
+  function _unstake(uint256 _index, uint256 _amount) internal override {
     _dHedgeSwapper.withdraw({
       pool: address(lpTokens[_index]),
       fundTokenAmount: _amount,
